+++ .gitignore
@@ -34,5 +34,9 @@ target/
 # Generated during runtime
 /shmem
 
+/smelter.app
+/smelter/
+/smelter_*.tar.gz
+
 .direnv
 /result
