+++ Cargo.toml
@@ -19,7 +19,7 @@ members = [
 
     # Tools/utils/tests
     "integration_tests",
-    "tools/generate",
+    "tools",
 ]
 resolver = "2"
 
@@ -45,6 +45,7 @@ crossbeam-channel = "0.5.15"
 reqwest = { version = "0.12.3", features = ["blocking", "json"] }
 signal-hook = "0.3.15"
 ffmpeg-next = "8.0.0"
+fs_extra = "1.3.0"
 anyhow = "1.0.71"
 image = { version = "0.24.7", features = ["jpeg", "png"] }
 itertools = "0.14.0"
@@ -86,7 +87,7 @@ ffmpeg-next = { workspace = true }
 crossbeam-channel = { workspace = true }
 log = { workspace = true }
 signal-hook = { workspace = true }
-fs_extra = "1.3.0"
+fs_extra = { workspace = true } 
 schemars = { workspace = true }
 image = { workspace = true }
 thiserror = { workspace = true }
