+++ compositor_chromium/Cargo.toml
@@ -8,7 +8,7 @@ license-file = "../LICENSE"
 
 [dependencies]
 chromium_sys = { path = "chromium_sys" }
-fs_extra = "1.3.0"
+fs_extra = { workspace = true }
 thiserror = { workspace = true }
 widestring = "1.0.2"
 log = { workspace = true }
