--- tools/generate/.gitignore
@@ -1,1 +1,0 @@
-/workingdir
