+++ tools/src/bin/generate_docs_examples/component_shader.wgsl
