+++ tools/src/bin/generate_docs_examples/component_text.rs
@@ -1,13 +1,13 @@
 use std::{fs, path::Path, process::Command};
 
 use anyhow::Result;
-use generate::compositor_instance::CompositorInstance;
 use serde_json::json;
+use tools::compositor_instance::CompositorInstance;
 
-pub(super) fn generate_view_component_example(root_dir: &Path) -> Result<()> {
+pub(super) fn generate_text_component_example(root_dir: &Path) -> Result<()> {
     let instance = CompositorInstance::start();
-    let mp4_path = root_dir.join("guides/component-view-example.mp4");
-    let webp_path = root_dir.join("guides/component-view-example.webp");
+    let mp4_path = root_dir.join("guides/component-text-example.mp4");
+    let webp_path = root_dir.join("guides/component-text-example.webp");
     let _ = fs::remove_file(&mp4_path);
 
     instance.send_request(
@@ -58,16 +58,30 @@ fn scene() -> serde_json::Value {
     json!({
         "root": {
             "type": "view",
+            "background_color": "#52505b",
+            "padding": 100,
+            "direction": "column",
             "children": [
                 {
-                    "type": "view",
-                    "direction": "column",
-                    "children": [
-                        { "type": "view", "background_color": "red", "height": 200 },
-                        { "type": "view", "background_color": "blue" }
-                    ],
+                    "type": "text",
+                    "font_size": 72,
+                    "color": "#a5baf0",
+                    "weight": "bold",
+                    "text": "Example text"
                 },
-                { "type": "view", "background_color": "green" }
+                {
+                    "type": "view",
+                    "height": 30,
+                },
+                {
+                    "type": "text",
+                    "font_size": 30,
+                    "line_height": 44,
+                    "wrap": "word",
+                    "width": 1000,
+                    "color": "#a5baf0",
+                    "text": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam consequat lorem a quam bibendum, non gravida tortor ornare. Cras blandit facilisis erat. Integer porta ullamcorper mauris ac maximus. Donec sapien diam, porttitor nec interdum sit amet, eleifend at lectus."
+                },
             ]
         }
     })
