+++ tools/src/bin/generate_docs_examples/main.rs
