+++ tools/src/bin/generate_docs_examples/transition.rs
@@ -1,12 +1,12 @@
 use std::{fs, path::Path};
 
 use anyhow::Result;
-use generate::compositor_instance::CompositorInstance;
 use serde_json::json;
+use tools::compositor_instance::CompositorInstance;
 
 use crate::workingdir;
 
-pub(super) fn generate_tile_component_example(root_dir: &Path) -> Result<()> {
+pub(super) fn generate_tile_transition_video(root_dir: &Path) -> Result<()> {
     let instance = CompositorInstance::start();
     let mp4_path = root_dir.join("guides/component-tiles-example.mp4");
     let _ = fs::remove_file(&mp4_path);
@@ -55,7 +55,7 @@ pub(super) fn generate_tile_component_example(root_dir: &Path) -> Result<()> {
             "type": "mp4",
             "path": workingdir().join("input_3.mp4").to_str().unwrap(),
             "required": true,
-            "offset_ms": 3000
+            "offset_ms": 0
         }),
     )?;
 
@@ -65,7 +65,7 @@ pub(super) fn generate_tile_component_example(root_dir: &Path) -> Result<()> {
             "type": "mp4",
             "path": workingdir().join("input_4.mp4").to_str().unwrap(),
             "required": true,
-            "offset_ms": 6000
+            "offset_ms": 0
         }),
     )?;
 
@@ -75,14 +75,14 @@ pub(super) fn generate_tile_component_example(root_dir: &Path) -> Result<()> {
             "type": "mp4",
             "path": workingdir().join("input_5.mp4").to_str().unwrap(),
             "required": true,
-            "offset_ms": 6000
+            "offset_ms": 0
         }),
     )?;
 
     instance.send_request(
         "output/output_1/unregister",
         json!({
-            "schedule_time_ms": 15_000,
+            "schedule_time_ms": 14_000,
         }),
     )?;
 
@@ -90,7 +90,15 @@ pub(super) fn generate_tile_component_example(root_dir: &Path) -> Result<()> {
         "output/output_1/update",
         json!({
             "video": scene(vec!["input_1", "input_2", "input_3"]),
-            "schedule_time_ms": 3000
+            "schedule_time_ms": 2000
+        }),
+    )?;
+
+    instance.send_request(
+        "output/output_1/update",
+        json!({
+            "video": scene(vec!["input_1", "input_2", "input_3", "input_4"]),
+            "schedule_time_ms": 4000
         }),
     )?;
 
@@ -105,15 +113,23 @@ pub(super) fn generate_tile_component_example(root_dir: &Path) -> Result<()> {
     instance.send_request(
         "output/output_1/update",
         json!({
-            "video": scene(vec!["input_1", "input_3", "input_4", "input_5"]),
-            "schedule_time_ms": 9_000
+            "video": scene(vec!["input_1", "input_2", "input_3", "input_5", "input_4"]),
+            "schedule_time_ms": 8000
         }),
     )?;
 
     instance.send_request(
         "output/output_1/update",
         json!({
-            "video": scene(vec!["input_1", "input_4", "input_5"]),
+            "video": scene(vec!["input_1", "input_2", "input_4"]),
+            "schedule_time_ms": 10_000
+        }),
+    )?;
+
+    instance.send_request(
+        "output/output_1/update",
+        json!({
+            "video": scene(vec!["input_1", "input_2"]),
             "schedule_time_ms": 12_000
         }),
     )?;
@@ -143,7 +159,7 @@ fn scene(inputs: Vec<&str>) -> serde_json::Value {
             "margin": 20,
             "background_color": "#52505bff",
             "transition": {
-                "duration_ms": 300,
+                "duration_ms": 500,
                 "easing_function": {
                     "function_name": "cubic_bezier",
                     "points": [0.35, 0.22, 0.1, 0.8]
