+++ tools/src/bin/generate_docs_examples/view_transitions.rs
@@ -4,8 +4,8 @@ use std::{
 };
 
 use anyhow::Result;
-use generate::compositor_instance::CompositorInstance;
 use serde_json::json;
+use tools::compositor_instance::CompositorInstance;
 
 use crate::workingdir;
