+++ tools/src/bin/package_for_release/bundle_linux.rs
@@ -1,11 +1,11 @@
 use anyhow::{anyhow, Result};
 use fs_extra::dir::{self, CopyOptions};
 use log::info;
-use std::fs;
-use std::path::PathBuf;
 use std::process::Command;
+use std::{fs, path::Path};
+use tools::paths::{git_root, tools_root};
 
-use crate::utils;
+use crate::utils::{self, SmelterBin};
 
 const X86_TARGET: &str = "x86_64-unknown-linux-gnu";
 const X86_OUTPUT_FILE: &str = "smelter_linux_x86_64.tar.gz";
@@ -17,16 +17,25 @@ const ARM_OUTPUT_FILE: &str = "smelter_linux_aarch64.tar.gz";
 pub fn bundle_linux_app() -> Result<()> {
     tracing_subscriber::fmt().init();
 
+    let workdir = tools_root().join("build");
+    utils::ensure_empty_dir(&workdir)?;
+
     if cfg!(target_arch = "x86_64") {
-        bundle_app(X86_TARGET, X86_OUTPUT_FILE, false)?;
-        bundle_app(X86_TARGET, X86_WITH_WEB_RENDERER_OUTPUT_FILE, true)?;
+        bundle_app(&workdir, X86_TARGET, X86_OUTPUT_FILE, false)?;
+        bundle_app(
+            &workdir,
+            X86_TARGET,
+            X86_WITH_WEB_RENDERER_OUTPUT_FILE,
+            true,
+        )?;
     } else if cfg!(target_arch = "aarch64") {
-        bundle_app(ARM_TARGET, ARM_OUTPUT_FILE, false)?;
+        bundle_app(&workdir, ARM_TARGET, ARM_OUTPUT_FILE, false)?;
     }
     Ok(())
 }
 
 fn bundle_app(
+    workdir: &Path,
     target_name: &'static str,
     output_name: &str,
     enable_web_rendering: bool,
@@ -37,55 +46,57 @@ fn bundle_app(
         info!("Bundling smelter without web rendering");
     }
 
-    let root_dir_str = env!("CARGO_MANIFEST_DIR");
-    let root_dir: PathBuf = root_dir_str.into();
-    let release_dir = root_dir.join(format!("target/{target_name}/release"));
-    let tmp_dir = root_dir.join("smelter");
-    utils::setup_bundle_dir(&tmp_dir)?;
+    let cargo_build_dir = git_root().join("target").join(target_name).join("release");
+    utils::ensure_empty_dir(&workdir.join("smelter"))?;
 
     info!("Build main_process binary.");
-    utils::cargo_build("main_process", target_name, !enable_web_rendering)?;
+    utils::compile_smelter(SmelterBin::MainProcess, target_name, !enable_web_rendering)?;
 
     if enable_web_rendering {
         info!("Build process_helper binary.");
-        utils::cargo_build("process_helper", target_name, false)?;
-
-        info!("Create {} directory", tmp_dir.display());
-        fs::create_dir_all(tmp_dir.clone())?;
+        utils::compile_smelter(SmelterBin::ChromiumHelper, target_name, false)?;
 
         info!("Copy main_process binary.");
         fs::copy(
-            release_dir.join("main_process"),
-            tmp_dir.join("smelter_main"),
+            cargo_build_dir.join("main_process"),
+            workdir.join("smelter/smelter_main"),
         )?;
 
         info!("Copy process_helper binary.");
         fs::copy(
-            release_dir.join("process_helper"),
-            tmp_dir.join("smelter_process_helper"),
+            cargo_build_dir.join("process_helper"),
+            workdir.join("smelter/smelter_process_helper"),
         )?;
 
         info!("Copy wrapper script.");
         fs::copy(
-            root_dir.join("src/bin/package_for_release/linux_runtime_wrapper.sh"),
-            tmp_dir.join("smelter"),
+            tools_root().join("src/bin/package_for_release/linux_runtime_wrapper.sh"),
+            workdir.join("smelter/smelter"),
         )?;
 
         info!(
             "Copy lib directory. {:?} {:?}",
-            release_dir.join("lib"),
-            tmp_dir.join("lib"),
+            cargo_build_dir.join("lib"),
+            workdir.join("smelter/lib"),
         );
 
-        dir::copy(release_dir.join("lib"), tmp_dir, &CopyOptions::default())?;
+        dir::copy(
+            cargo_build_dir.join("lib"),
+            workdir.join("smelter"),
+            &CopyOptions::default(),
+        )?;
     } else {
         info!("Copy main_process binary.");
-        fs::copy(release_dir.join("main_process"), tmp_dir.join("smelter"))?;
+        fs::copy(
+            cargo_build_dir.join("main_process"),
+            workdir.join("smelter/smelter"),
+        )?;
     }
 
     info!("Create tar.gz archive.");
     let exit_code = Command::new("tar")
-        .args(["-C", root_dir_str, "-czvf", output_name, "smelter"])
+        .args(["-czvf", output_name, "smelter"])
+        .current_dir(workdir)
         .spawn()?
         .wait()?
         .code();
