+++ tools/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod compositor_instance;
+pub mod paths;
